moved isAuthenticated and isAdmin to AuthenticationUtils
